backend/requirements.txt
azure-ai-projects
python-dotenv
azure-identity
fastapi
uvicorn
python-jose